+++ src/lid_driven_cavity/solver.rs
@@ -5,10 +5,10 @@ use super::{LidDrivenCavity, Links};
 impl LidDrivenCavity {
     pub fn solver(
         &self,
-        x: &mut [f64],
-        a_0: &[f64],
+        x: &mut [f32],
+        a_0: &[f32],
         links: &[Links],
-        sources: &[f64],
+        sources: &[f32],
         iter: usize,
     ) {
         let n = self.nx;
@@ -226,14 +226,14 @@ impl LidDrivenCavity {
 }
 
 #[inline(always)]
-fn replace_row(row: usize, x: &mut [f64], solx: &[f64], nx: usize) {
+fn replace_row(row: usize, x: &mut [f32], solx: &[f32], nx: usize) {
     for i in 0..nx {
         x[ix(i, row, nx)] = solx[i];
     }
 }
 
 #[inline(always)]
-fn replace_column(column: usize, x: &mut [f64], soly: &[f64], nx: usize, ny: usize) {
+fn replace_column(column: usize, x: &mut [f32], soly: &[f32], nx: usize, ny: usize) {
     for j in 0..ny {
         x[ix(column, j, nx)] = soly[j];
     }
